tests/requirements_test.txt
coverage==7.1.0
croniter==1.3.8
watchdog==2.1.9
mock-open==1.4.0
mypy==1.1.1
pre-commit==3.2.1
pytest==7.2.1
pytest-cov==3.0.0
pytest-homeassistant-custom-component==0.13.10
pylint==2.17.1
pylint-strict-informational==0.1
